+++ test/fuzz-corpus.sh
@@ -23,7 +23,7 @@ set -u
 rc=0
 found=0
 
-for h in config pim igmp; do
+for h in config pim igmp ipc; do
 	bin="./fuzz_${h}_replay"
 	dir="$srcdir/fuzz/corpus/$h"